requirements.txt
algoliasearch==1.15.0
algoliasearch-django==1.5.1
amqp==2.2.2
asn1crypto==0.24.0
billiard==3.5.0.3
bintrees==2.0.7
boto==2.48.0
boto3==1.4.4
Celery==4.1.0
chardet==3.0.4
certifi==2017.7.27.1
cffi==1.11.5
crc16==0.1.1
cryptography==2.3
dateparser==0.7.0
Django==1.11.13
django-allauth==0.36.0
django-betterforms==1.1.4
django-bootstrap3==10.0.1
django-bootstrap-form==3.4
django-bower==5.2.0
django-bulk-update==2.2.0
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-storages==1.6.5
django-timeseries==1.0.2
django-webpack-loader==0.6.0
djangorestframework-jwt==1.11.0
djangorestframework==3.6.3
ed25519==1.4
enum34==1.1.6
gdax==1.0.6
idna==2.7
ipaddress==1.0.22
kombu==4.1.0
mnemonic==0.18
oauthlib==2.0.7
olefile==0.44
Pillow==4.3.0
parse==1.8.4
pbkdf2==1.3
pycparser==2.18
pycryptodomex==3.4.7
pyjwt==1.3.0
pyOpenSSL==17.0.0
python-binance==0.6.9
python-magic==0.4.13
python-openid==2.2.5
pytz==2018.4
regex==2018.2.21
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
SSEClient==0.0.19
stellar-base==0.1.8.1
stream-django==1.4.0
stream-python==2.11.0
toml==0.9.4
tzlocal==1.5.1
urllib3==1.22
vine==1.1.4
